refactor: drop duplication of media validation
