+++ src/pages/Howto/labels.ts
@@ -6,6 +6,8 @@ import {
   HOWTO_TITLE_MAX_LENGTH,
 } from './constants'
 
+import { videoUrl } from '../../../src/utils/labels'
+
 export const headings = {
   create: '<span>Create</span> a How-To',
   edit: '<span>Edit</span> a How-To',
@@ -88,13 +90,5 @@ export const steps = {
     placeholder: `Explain what you are doing. If it gets too long, consider breaking it into multiple steps (${HOWTO_STEP_DESCRIPTION_MIN_LENGTH}-${HOWTO_STEP_DESCRIPTION_MAX_LENGTH} characters)`,
   },
   images: 'Upload image(s)',
-  videoUrl: {
-    title: 'Or embed a YouTube video',
-    placeholder: 'https://youtube.com/watch?v=',
-    errors: {
-      both: 'Do not include both images and video',
-      empty: 'Include either images or a video',
-      invalidUrl: 'Please provide a valid YouTube Url',
-    },
-  },
+  videoUrl,
 }
