+++ src/utils/labels.ts
@@ -1,0 +1,9 @@
+export const videoUrl = {
+    errors: {
+      both: 'Do not include both images and video',
+      empty: 'Include either images or a video',
+      invalidUrl: 'Please provide a valid YouTube Url',
+    },
+    title: 'Or embed a YouTube video',
+    placeholder: 'https://youtube.com/watch?v=',
+}
